+++ Controllers/ProjectsController.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -14,12 +15,10 @@ namespace Review.Api.Controllers;
 public class ProjectsController : ControllerBase
 {
     private readonly IProjectService _service;
-    private readonly UserManager<User> _userManager;
 
-    public ProjectsController(UserManager<User> userManager, IProjectService service)
+    public ProjectsController(IProjectService service)
     {
         _service = service;
-        _userManager = userManager;
     }
 
     [HttpGet]
@@ -29,9 +28,10 @@ public class ProjectsController : ControllerBase
         // TODO implement global exception handler
         try
         {
-            var currentUser = await _userManager.GetUserAsync(User);
-            
-            var result = await _service.GetAllAsync();
+            // var currentUser = await _userManager.GetUserAsync(User);
+            string? userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (string.IsNullOrEmpty(userId)) return Unauthorized();
+            var result = await _service.GetAllAsync(userId);
             return Ok(result);
         }
         catch (Exception e)
@@ -42,15 +42,17 @@ public class ProjectsController : ControllerBase
     }
 
     [HttpGet]
-    [Route("{id}")]
-    public async Task<ActionResult<ProjectWithAssetsDto>> GetById(string id)
+    [Route("{projectId}")]
+    public async Task<ActionResult<ProjectWithAssetsDto>> GetById(string projectId)
     {
         // TODO Validate incoming API key before returning project details.
         // TODO implement global exception handler
         try
         {
-            ProjectWithAssetsDto? result = await _service.GetByIdAsync(id);
-            if (result == null) return NotFound($"Project {id} not found");
+            string? userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (string.IsNullOrEmpty(userId)) return Unauthorized();
+            ProjectWithAssetsDto? result = await _service.GetByIdAsync(userId, projectId);
+            if (result == null) return NotFound($"Project {projectId} not found");
             return Ok(result);
         }
         catch (Exception e)
@@ -70,9 +72,9 @@ public class ProjectsController : ControllerBase
         try
         {
             if (!ModelState.IsValid) return ValidationProblem(ModelState);
-            var currentUser = await _userManager.GetUserAsync(User);
-            if (currentUser == null) return Unauthorized();
-            var result = await _service.CreateAsync(currentUser.Id, data);
+            string? userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (string.IsNullOrEmpty(userId)) return Unauthorized();
+            var result = await _service.CreateAsync(userId, data);
             // if (result == null) return BadRequest("Project couldn't be created");
 
             return CreatedAtAction(nameof(GetById), new { id = result.Id }, result);
@@ -85,16 +87,18 @@ public class ProjectsController : ControllerBase
     }
 
     [HttpPut]
-    [Route("{id}")]
-    public async Task<ActionResult> Update(string id, Project project)
+    [Route("{projectId}")]
+    public async Task<ActionResult> Update(string projectId, Project project)
     {
         // TODO Validate incoming API key before updating a project.
         // TODO implement global exception handler
         // TODO validate data
         try
         {
-            var result = await _service.UpdateAsync(id, project);
-            if (!result) return BadRequest($"Couldn't update project {id}");
+            string? userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (string.IsNullOrEmpty(userId)) return Unauthorized();
+            var result = await _service.UpdateAsync(userId, projectId, project);
+            if (!result) return BadRequest($"Couldn't update project {projectId}");
             return NoContent();
         }
         catch (Exception e)
@@ -105,15 +109,17 @@ public class ProjectsController : ControllerBase
     }
 
     [HttpDelete]
-    [Route("{id}")]
-    public async Task<ActionResult> Delete(string id)
+    [Route("{projectId}")]
+    public async Task<ActionResult> Delete(string projectId)
     {
         // TODO Validate incoming API key before deleting a project.
         // TODO implement global exception handler
         try
         {
-            var result = await _service.DeleteAsync(id);
-            if (!result) return BadRequest($"Couldn't delete project {id}");
+            string? userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (string.IsNullOrEmpty(userId)) return Unauthorized();
+            var result = await _service.DeleteAsync(userId, projectId);
+            if (!result) return BadRequest($"Couldn't delete project {projectId}");
             return NoContent();
         }
         catch (Exception e)