+++ Program.cs
@@ -12,12 +12,15 @@ builder.Services.AddScoped<IUserRepository, UserRepository>();
 builder.Services.AddScoped<IProjectRepository, ProjectRepository>();
 builder.Services.AddScoped<IProjectService, ProjectService>();
 
-builder.Services.AddAutoMapper(options => { options.AddMaps(typeof(Program)); });
+builder.Services.AddAutoMapper(options =>
+{
+    options.LicenseKey = builder.Configuration["AutoMapper:ApiKey"];
+    options.AddMaps(typeof(Program));
+});
 
 const string databaseName = "ReviewDatabase";
 
-builder.Services.AddDbContext<AppDbContext>(options => options.UseInMemoryDatabase(databaseName));
-builder.Services.AddDbContext<IdentityContext>(options => options.UseInMemoryDatabase(databaseName));
+builder.Services.AddDbContext<ApplicationDbContext>(options => options.UseInMemoryDatabase(databaseName));
 
 // Add services to the container.
 builder.Services.AddControllers().AddJsonOptions(options =>
@@ -46,7 +49,7 @@ builder.Services.AddIdentityApiEndpoints<User>(options =>
     // TODO set owasp password standards
 
     options.User.RequireUniqueEmail = true;
-}).AddEntityFrameworkStores<IdentityContext>();
+}).AddEntityFrameworkStores<ApplicationDbContext>();
 
 builder.Services.AddAuthorization();
 
@@ -54,11 +57,8 @@ var app = builder.Build();
 
 using (var scope = app.Services.CreateScope())
 {
-    var context = scope.ServiceProvider.GetRequiredService<AppDbContext>();
+    var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
     context.Database.EnsureCreated();
-
-    var identityContext = scope.ServiceProvider.GetRequiredService<IdentityContext>();
-    identityContext.Database.EnsureCreated();
 
 
     // TODO remove before deployment