+++ Repositories/UserRepository.cs
@@ -4,11 +4,10 @@ using Review.Api.Models;
 
 namespace Review.Api.Repositories;
 
-public class UserRepository(AppDbContext context) : IUserRepository
+public class UserRepository(ApplicationDbContext context) : IUserRepository
 {
-    private readonly AppDbContext _context = context;
-
-
+    private readonly ApplicationDbContext _context = context;
+    
     public async Task<User?> GetByIdAsync(string id) =>
         await _context.Users.FirstOrDefaultAsync(p => p.Id == id);